requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
apache-airflow-providers-postgres==5.6.0
python-dotenv==1.0.1
